capstoneProject/cap5project/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
